+++ dll/steam_apps.cpp
@@ -93,7 +93,11 @@ bool Steam_Apps::BIsDlcInstalled( AppId_t appID )
     std::lock_guard<std::recursive_mutex> lock(global_mutex);
     if (appID == 0) return true;
     if (appID == UINT32_MAX) return false; // check Steam_Apps::BIsAppInstalled()
-    if (appID == settings->get_local_game_id().AppID()) return true; //TODO is this correct?
+    
+    // Age of Empires 2: Definitive Edition expects the app itself to be an owned DLC.
+    // otherwise it will only load the "Return of Rome" game mode, also most options are disabled
+    if (appID == settings->get_local_game_id().AppID()) return true;
+
     return settings->hasDLC(appID);
 }