+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* **[bitsynth]** Fix Age of Empires 2: Definitive Edition, the game expects the app itself to be an owned DLC  
+  otherwise most options will be disabled
+
+---
+
 # 2024/3/16
 * manage overlay cursor input/clipping and internal frame processing in a better way,
   should prevent more games from pausing to display notifications